requirements.txt
ultralytics
opencv-python-headless
numpy
pandas
scikit-learn
joblib
matplotlib
pillow
streamlit
requests
tqdm